기능 변경 : 채팅방 삭제 기능 변경
채팅방 삭제 시 해당 채팅방의 참여 유저 목록도 함께 제거하도록 변경

+++ src/main/java/com/noti/chatapp/controller/chat/ChatRoomController.java
@@ -130,8 +130,8 @@ public class ChatRoomController {
 
     @DeleteMapping(value = "/api/chat/delete-room")
     public @ResponseBody
-    ResponseEntity<Long> chatRoomDelete(@Valid @RequestBody ChatRoomDto requestDto) {
-        long result = chatRoomService.deleteByRoomIdAndRoomPw(requestDto);
+    ResponseEntity<Integer> chatRoomDelete(@Valid @RequestBody ChatRoomDto requestDto) {
+        int result = chatRoomService.deleteByRoomIdAndRoomPw(requestDto);
         return ResponseEntity.ok(result);
     }
